[AXON-1545] Move rovodev analytics logic
