+++ src/rovo-dev/api/extensionApi.test.ts
@@ -14,7 +14,6 @@ const mockSiteManager = {
 // Mock Container
 jest.mock('src/container', () => ({
     Container: {
-        analyticsClient: { sendTrackEvent: jest.fn() },
         isDebugging: false,
         isBoysenberryMode: false,
         isRovoDevEnabled: true,
@@ -57,15 +56,6 @@ describe('ExtensionApi', () => {
     beforeEach(() => {
         jest.clearAllMocks();
         api = new ExtensionApi();
-    });
-
-    describe('analytics', () => {
-        it('should send track event through Container', async () => {
-            const mockEvent = { type: 'test', data: 'value' };
-            await api.analytics.sendTrackEvent(mockEvent);
-
-            expect(Container.analyticsClient.sendTrackEvent).toHaveBeenCalledWith(mockEvent);
-        });
     });
 
     describe('metadata', () => {
