+++ src/rovo-dev/performanceLogger.ts
@@ -1,5 +1,5 @@
-import { performanceEvent, RovoDevEnv } from '../../src/analytics';
 import Perf from '../util/perf';
+import { RovoDevEnv } from './analytics/rovodevAnalyticsTypes';
 import { ExtensionApi } from './api/extensionApi';
 import { RovoDevLogger } from './util/rovoDevLogger';
 
@@ -26,12 +26,16 @@ export class PerformanceLogger {
 
     public async promptFirstByteReceived(promptId: string) {
         const measure = Perf.measure(promptId);
-        const evt = await performanceEvent('api.rovodev.chat.response.timeToFirstByte', measure, {
-            rovoDevEnv: this.rovoDevEnv,
-            appInstanceId: this.appInstanceId,
-            rovoDevSessionId: this.currentSessionId,
-            rovoDevPromptId: promptId,
-        });
+        const evt = await this.extensionApi.analytics.performanceEvent(
+            'api.rovodev.chat.response.timeToFirstByte',
+            measure,
+            {
+                rovoDevEnv: this.rovoDevEnv,
+                appInstanceId: this.appInstanceId,
+                rovoDevSessionId: this.currentSessionId,
+                rovoDevPromptId: promptId,
+            },
+        );
 
         RovoDevLogger.debug(`Event fired: rovodev.response.timeToFirstByte ${measure} ms`);
         await this.extensionApi.analytics.sendTrackEvent(evt);
@@ -39,12 +43,16 @@ export class PerformanceLogger {
 
     public async promptFirstMessageReceived(promptId: string) {
         const measure = Perf.measure(promptId);
-        const evt = await performanceEvent('api.rovodev.chat.response.timeToFirstMessage', measure, {
-            rovoDevEnv: this.rovoDevEnv,
-            appInstanceId: this.appInstanceId,
-            rovoDevSessionId: this.currentSessionId,
-            rovoDevPromptId: promptId,
-        });
+        const evt = await this.extensionApi.analytics.performanceEvent(
+            'api.rovodev.chat.response.timeToFirstMessage',
+            measure,
+            {
+                rovoDevEnv: this.rovoDevEnv,
+                appInstanceId: this.appInstanceId,
+                rovoDevSessionId: this.currentSessionId,
+                rovoDevPromptId: promptId,
+            },
+        );
 
         RovoDevLogger.debug(`Event fired: rovodev.response.timeToFirstMessage ${measure} ms`);
         await this.extensionApi.analytics.sendTrackEvent(evt);
@@ -52,12 +60,16 @@ export class PerformanceLogger {
 
     public async promptTechnicalPlanReceived(promptId: string) {
         const measure = Perf.measure(promptId);
-        const evt = await performanceEvent('api.rovodev.chat.response.timeToTechPlan', measure, {
-            rovoDevEnv: this.rovoDevEnv,
-            appInstanceId: this.appInstanceId,
-            rovoDevSessionId: this.currentSessionId,
-            rovoDevPromptId: promptId,
-        });
+        const evt = await this.extensionApi.analytics.performanceEvent(
+            'api.rovodev.chat.response.timeToTechPlan',
+            measure,
+            {
+                rovoDevEnv: this.rovoDevEnv,
+                appInstanceId: this.appInstanceId,
+                rovoDevSessionId: this.currentSessionId,
+                rovoDevPromptId: promptId,
+            },
+        );
 
         RovoDevLogger.debug(`Event fired: rovodev.response.timeToTechPlan ${measure} ms`);
         await this.extensionApi.analytics.sendTrackEvent(evt);
@@ -65,12 +77,16 @@ export class PerformanceLogger {
 
     public async promptLastMessageReceived(promptId: string) {
         const measure = Perf.measure(promptId);
-        const evt = await performanceEvent('api.rovodev.chat.response.timeToLastMessage', measure, {
-            rovoDevEnv: this.rovoDevEnv,
-            appInstanceId: this.appInstanceId,
-            rovoDevSessionId: this.currentSessionId,
-            rovoDevPromptId: promptId,
-        });
+        const evt = await this.extensionApi.analytics.performanceEvent(
+            'api.rovodev.chat.response.timeToLastMessage',
+            measure,
+            {
+                rovoDevEnv: this.rovoDevEnv,
+                appInstanceId: this.appInstanceId,
+                rovoDevSessionId: this.currentSessionId,
+                rovoDevPromptId: promptId,
+            },
+        );
 
         Perf.clear(promptId);
